server/requirements.txt
Flask==2.3.2
Flask-CORS==3.0.10
python-dotenv==1.0.0
boto3 